Move Liquid event handlers in "/Services" folders.
